+++ docs/src/components/Sidebar/Sidebar.tsx
@@ -297,6 +297,9 @@ class Sidebar extends React.Component<any, any> {
                   <Menu.Item as={NavLink} exact to="/prototype-popups" activeClassName="active">
                     Popups
                   </Menu.Item>
+                  <Menu.Item as={NavLink} exact to="/icon-viewer" activeClassName="active">
+                    Processed Icons
+                  </Menu.Item>
                 </Menu.Menu>
               </Menu.Item>
             )}