+++ src/themes/teams/components/Icon/iconVariables.ts
@@ -3,6 +3,9 @@ import { pxToRem } from '../../../../lib'
 export interface IconVariables {
   [key: string]: string | number | boolean | undefined
 
+  redColor?: string
+  brandColor?: string
+
   outline?: boolean
   color?: string
   backgroundColor?: string
@@ -22,4 +25,7 @@ export default (siteVars): IconVariables => ({
   margin: `0 ${pxToRem(8)} 0 0`,
   secondaryColor: siteVars.white,
   disabledColor: siteVars.gray06,
+
+  redColor: siteVars.red,
+  brandColor: siteVars.brandColor,
 })